# api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: DBackup API
  version: 2.0.0
  description: |
    REST API for DBackup - a self-hosted database backup automation platform with encryption, compression, and smart retention.

    ## Authentication

    DBackup supports two authentication methods:

    - **Session Authentication** - Used automatically when logged in via the web UI. Session cookies are sent with each request.
    - **API Key Authentication** - For scripts, CI/CD pipelines, and external integrations. Create an API key under **Access Management → API Keys**.

    > API keys do not inherit SuperAdmin privileges. Only explicitly assigned permissions are available.

    ## Rate Limits

    | Request Type | Default Limit |
    |---|---|
    | Authentication (login, etc.) | 5/min per IP |
    | GET requests | 100/min per IP |
    | POST / PUT / DELETE | 20/min per IP |

    Rate limits are configurable in Settings → Rate Limits.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
...
servers:
  - url: "{protocol}://{host}/api"
    description: DBackup Instance
    variables:
      protocol:
        default: https
        enum: [http, https]
      host:
        default: your-dbackup-instance.com
security:
  - BearerAuth: []
  - SessionAuth: []
...
paths:
...
  /storage/{id}/files:
    get:
      tags: [Storage Explorer]
      summary: List Backup Files
      description: Lists all backup files in a storage destination with enriched metadata.
      operationId: listBackupFiles
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - name: typeFilter
          in: query
          schema:
            type: string
          description: Filter by file type
      responses:
        "200":
          description: Backup file list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Storage Explorer]
      summary: Delete Backup File
      operationId: deleteBackupFile
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  description: File path to delete
                  example: "backups/mydb_2025-01-15.sql.gz.enc"
      responses:
        "200":
          description: File deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /storage/{id}/history:
    get:
      tags: [Storage Explorer]
      summary: Storage Usage History
      description: Returns storage size and file count snapshots over time.
      operationId: getStorageHistory
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - name: days
          in: query
          schema:
            type: integer
            default: 30
            minimum: 1
            maximum: 365
          description: Number of days of history
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Storage history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        totalSize:
                          type: integer
                          description: Total size in bytes
                        fileCount:
                          type: integer
                        createdAt:
                          type: string
                          format: date-time

  # ─── Vault ─────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API Key authentication. Create an API key under **Access Management → API Keys**.

        Format: `dbackup_xxxxxxxx`
    SessionAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie (browser login)
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
  responses:
    Unauthorized:
      description: Missing, invalid, disabled, or expired credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Valid credentials but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Human-readable error message"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
    FileInfo:
      type: object
      properties:
        name:
          type: string
          example: "mydb_2025-01-15.sql.gz.enc"
        path:
          type: string
        size:
          type: integer
          description: Size in bytes
        lastModified:
          type: string
          format: date-time
        locked:
          type: boolean
          description: Whether the file is currently being written
...